Add filtering by location
Adds the ability to filter by latitude and longitude to look for addresses
at an exact location, or addresses within a distance from that location.
If no distance is given, 1 meter is used as the distance.

To speed up spatial lookups, we create a buffer and limit our search within
that buffer. Without the buffer, we would have to compare the distances of
all addresses in the database, which can take a few seconds. The buffer
makes this a lot faster.

For the buffer, we need some kind of approximation of the distance in
degrees, since WGS84 uses degrees as the unit. This is done by calculating
the degrees based on Earth's circumference (40,000,000 meters). It's an
approximation, so I have doubled the distance just in case to make sure we
don't make the buffer too tight.