Add ability to get responses in XML format
By default, the API returns JSON responses. If we browse the API with the
browser, the browsable API will be shown instead. With the `&format=xxx`
you can specify the format explicitly, so to get the data in XML format,
you can add `&format=xml` to the end of the URL.